pyproject.toml
[project]
name = "fraud-intel"
version = "0.1.0"
description = "DSPy-powered fraud intelligence and analysis system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy>=2.6.27",
    "openinference-instrumentation-dspy>=0.1.0",
    "openinference-instrumentation-litellm>=0.1.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 120
